mnist_project/src/app/requirements.txt
torch
torchvision
tqdm
fastapi
uvicorn
pydantic
Pillow
streamlit
streamlit-drawable-canvas
requests
numpy
pillow
opencv-python-headless